Implement local storage and fix bugs
Changing to returning a copy of the list of todo items
instead of the list of todo items broke some functionality
for updating the todo items. Everywhere that was using the
copied todo item was not persisting, so it was necessary
to get the real todo items before making changes.